+++ lib/archer/engine.rb
@@ -1,8 +1,11 @@
 module Archer
   class Engine < Rails::Engine
     console do
+      # TODO remove in 0.3.0
       Archer.history_file ||= Rails.root.join("tmp", ".irb-history")
-      Archer.start
+
+      # need IRB context to be ready before Archer.start
+      ::IRB::Irb.prepend(Archer::Irb)
 
       at_exit do
         Archer.save