+++ lib/archer/irb.rb
@@ -1,0 +1,8 @@
+module Archer
+  module Irb
+    def run(*)
+      Archer.start
+      super
+    end
+  end
+end